% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dhp.dist}
\alias{dhp.dist}
\title{Estimate diameter frequency distribution}
\usage{
dhp.dist(dbh, weibull_params)
}
\arguments{
\item{dbh}{numeric. vector of diameters at breast height (dbh) to run the simulation}

\item{weibull_param}{dataframe. output of function param_weibull_pct}
}
\value{
a dataframe with number of stem per hectare for each dbh value given as input
}
\description{
This function estimates the diameter frequency distribution in unthinned white spruce plantations using
the weibull function parameters 'location', 'scale' and 'shape'.
}
