% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{height.epb}
\alias{height.epb}
\title{Estimating tree height}
\usage{
height.epb(dbh, mdbh, MAT, SDOM, VPOT, BA)
}
\arguments{
\item{dbh}{numeric. diameter at breast height (cm)}

\item{mdbh}{numeric. mean dbh at breast height of the plantation (cm)}

\item{MAT}{numeric. mean annual temperature}

\item{SDOM}{numeric. coefficient associated with bioclimatic subdomain}

\item{VPOT}{numeric. coefficient associated with potential vegetation}

\item{BA}{numeric. merchantable basal area (m2/ha)}
}
\value{
a dataframe containing the height of each given dbh value
}
\description{
This function predicts the height of trees in unthinned white spruce plantations based on Auger (2016).
Values for parameters SDOM and VPOT must be extracted from https://mffp.gouv.qc.ca/publications/forets/connaissances/recherche/Auger-Isabelle/Note146.pdf
}
\examples{
height.epb(dbh = 10:20, 
           mdbh = 12, 
           MAT = 2.5, 
           SDOM = -0.2749, 
           VPOT = -0.1112, 
           BA = seq(20, 40, by = 2))
  
}
