% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{statSAW.plantation}
\alias{statSAW.plantation}
\title{StatSAW : Modelling lumber product assortment}
\usage{
statSAW.plantation(dbh, vol_dm3, stem_ha, h, return.volume = FALSE)
}
\arguments{
\item{dbh}{numeric. diameters at breast height (dbh)}

\item{stem_ha}{numeric. Number of stem per hectare}

\item{h}{numeric. (optional) Height(m) of trees of each dbh value}

\item{vol_m}{numeric. (optional) merchantable volume (dm3) for each given dbh value}
}
\value{
a dataframe with number and volume in pmp of lumber products for a diameter frequency distribution

The function uses the following six product categories
1x3x08 <- c("1x3x06", "1x3x07", "1x3x08", "1x3x10", "1x3x12", "1x3x14")
1x4x08 <- c("1x4x06", "1x4x07", "1x4x08", "1x4x10", "1x4x14", "1x4x16", "1x6x08")
2x3x08 <- c("2x3x05", "2x3x06", "2x3x07", "2x3x08")
2x4x08 <- c("2x4x05", "2x4x06", "2x4x07", "2x4x08")
2x4x16 <- c("2x4x10", "2x4x12", "2x4x14", "2x4x16")
2x6x10 <- c("2x6x07", "2x6x08", "2x6x10", "2x6x12", "2x6x14", "2x6x16")
}
\description{
Implement StatSAW model as described in Auty et al. 2014 (StatSAW: modelling lumber product assortment using zero-inflated Poisson regression,
dx.doi.org/10.1139/cjfr-2013-0500), which was refitted for white and black spruce plantation in 2021.
}
\details{
This function predicts the number and volume of lumber products that can be obtained in unthinned white spruce plantation given its dbh (diameter at breast height)
frequency distribution. If merchantable volume  (dm3) for each dbh valueis not given, it is calculated using the height and dbh or the trees using the method described in
"Équation provenant du mémoire de recherche no 160 (2010): Tarif de cubage, tables de rendement et modèles de croissance pour les plantations d'épinette blanche au Québec
https://mffp.gouv.qc.ca/publications/forets/connaissances/recherche/Pregent-Guy/Memoire160.pdf" which is fitted for plantations located in Quebec, Canada.
}
\examples{
param = param_weibull_pct(age = 23, iqs = 11.04, tige_ha = 1200, h = 8.29)
stem_per_ha = dhp_dist(1:60, param)

}
