% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weibull.param.pct}
\alias{weibull.param.pct}
\title{Estimate Weibull parameters to predict the diameter frequency distribution}
\usage{
weibull.param.pct(age, iqs, stem_ha, h, mdbh, id)
}
\arguments{
\item{age}{numeric. Age of plantation}

\item{iqs}{numeric. IQS (Station quality index) indicator of the productivity of a given site, based on the
mean height of the dominant tree in a a 50-yrs-old plantation.}

\item{stem_ha}{numeric. Number of stem per hectare}

\item{h}{numeric. Mean height of the dominant tree in the plantation}

\item{mdbh}{numeric. Arithmetic mean diameter at breast height of trees.}

\item{id}{character. Unique id for each plantation}
}
\value{
a dataframe with 5 columns named 'id', location', 'shape', 'scale' and 'stem_ha'

corresponding to the id of each plantation, the three parameters used to predict the distribution
of the diameters in a white spruce plantation, and the stem density
}
\description{
This function predicts the parameters of the Weibull distribution used to predict the diameter frequency distribution in unthinned white spruce plantations located in Quebec, Canada
based on the percentile-based parameter recovery approach (PCT) as described in Liu et al. 2009
(https://doi.org/10.1080/02827580802644599)
}
\examples{
weibull_param_pct(age = 23, iqs = 11.04, tige_ha = 1200, h = 8.29)

}
