% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{weibull.params.pct}
\alias{weibull.params.pct}
\title{Estimate Weibull parameters to predict the diameter frequency distribution}
\usage{
weibull.params.pct(id, age, iqs, stem.ha, h, mdbh)
}
\arguments{
\item{id}{character. unique id for each plantation}

\item{age}{numeric. age of plantation (years)}

\item{iqs}{numeric. station quality index, indicator of the productivity of a given site, based on the
mean height of the dominant tree in a a 50-yrs-old plantation.}

\item{stem.ha}{numeric. number of stems per hectare}

\item{h}{numeric. mean height of the dominant tree in the plantation (m)}

\item{mdbh}{numeric. mean diameter at breast height of trees (cm)}
}
\value{
a dataframe with 5 columns named 'id', location', 'shape', 'scale' and 'stem.ha'

corresponding to the id of each plantation, the three parameters used to predict the distribution
of the diameters in a white spruce plantation, and the stem density
}
\description{
This function predicts the parameters of the Weibull distribution used to predict the diameter frequency distribution in unthinned white spruce plantations located in Quebec, Canada
based on the percentile-based parameter recovery approach (PCT) as described in Liu et al. 2009
(https://doi.org/10.1080/02827580802644599)
}
\examples{
weibull.params.pct(id = "ID", 
                   age = 23, 
                   iqs = 11.04, 
                   stem.ha = 1200, 
                   h = 8.29, 
                   mdbh = 12)

}
