% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{height.epb}
\alias{height.epb}
\title{Estimating tree height}
\usage{
height.epb(dbh, mdbh, MAT, SDOM, VPOT, BA)
}
\arguments{
\item{dbh}{numeric.}

\item{mdbh}{numeric.}

\item{MAT}{numeric. Mean annual temperature}

\item{SDOM}{numeric. Coefficient associated with bioclimatic subdomain. See https://mffp.gouv.qc.ca/publications/forets/connaissances/recherche/Auger-Isabelle/Note146.pdf}

\item{VPOT}{numeric. Coefficient associated with potential vegetation. See https://mffp.gouv.qc.ca/publications/forets/connaissances/recherche/Auger-Isabelle/Note146.pdf}

\item{BA}{numeric. Merchantable basal area (m2/ha)}
}
\value{
a dataframe containing the height of each given dbh value
}
\description{
This function predicts the height of trees in unthinned white spruce plantations based on Auger 2016
(Une nouvelle relation hauteur-diamètre tenant compte de l’influence de la station et du climat pour 27 essences commerciales du Québec)
https://mffp.gouv.qc.ca/publications/forets/connaissances/recherche/Auger-Isabelle/Note146.pdf
}
