% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{statSAW.plantation}
\alias{statSAW.plantation}
\title{StatSAW : Modelling lumber product assortment}
\usage{
statSAW.plantation(id, dbh, vol.dm3, stem.ha, h, return.volume = FALSE)
}
\arguments{
\item{id}{character. unique id for each plantation}

\item{dbh}{numeric. diameters at breast height (cm)}

\item{vol.dm3}{numeric. (optional) merchantable volume (dm3) for each given dbh value}

\item{stem.ha}{numeric. number of stem per hectare}

\item{h}{numeric. (optional) height(m) of trees of each dbh value}

\item{return.volume}{logical. should the function output include the calculated merchantable volume (dm3)}
}
\value{
a dataframe with number and volume in pmp of lumber products for a single tree of given dbh

The function uses the following six product categories
\enumerate{
\item 1x3x08 including 1x3x06, 1x3x07, 1x3x08, 1x3x10, 1x3x12, 1x3x14
\item 1x4x08 including 1x4x06, 1x4x07, 1x4x08, 1x4x10, 1x4x14, 1x4x16, 1x6x08
\item 2x3x08 including 2x3x05, 2x3x06, 2x3x07, 2x3x08
\item 2x4x08 including 2x4x05, 2x4x06, 2x4x07, 2x4x08
\item 2x4x16 including 2x4x10, 2x4x12, 2x4x14, 2x4x16
\item 2x6x10 including 2x6x07, 2x6x08, 2x6x10, 2x6x12, 2x6x14, 2x6x16
}
}
\description{
Implement StatSAW model as described in Auty et al. 2014 (StatSAW: modelling lumber product assortment using zero-inflated Poisson regression,
dx.doi.org/10.1139/cjfr-2013-0500), which was refitted for white and black spruce plantation in 2021.
}
\details{
This function predicts the number and volume of lumber products that can be obtained in unthinned white spruce plantation given its dbh (diameter at breast height, cm)
frequency distribution. If merchantable volume  (dm3) for each dbh valueis not given, it is calculated using the height and dbh or the trees using the method described in
"Équation provenant du mémoire de recherche no 160 (2010): Tarif de cubage, tables de rendement et modèles de croissance pour les plantations d'épinette blanche au Québec
https://mffp.gouv.qc.ca/publications/forets/connaissances/recherche/Pregent-Guy/Memoire160.pdf" which is fitted for plantations located in Quebec, Canada.
}
\examples{
statSAW.plantation(id = as.character(rep(10:20, each = 10)),
                   dbh = 10:20, 
                   stem.ha = seq(10,100, by = 10), 
                   h = seq(10,15, by = 0.5),
                   return.volume = FALSE)

}
