% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{statSAW.value}
\alias{statSAW.value}
\title{Estimating the monetary value of lumber product assortments}
\usage{
statSAW.value(lumber, value)
}
\arguments{
\item{lumber}{lumber product assortment. output of function statSAW.plantation().}

\item{value}{dataframe comprising two columns; 1. category of lumber product (as in lumber) and 2. associated value}
}
\description{
This function estimate the monetary value of with lumber product assortments
}
